compas/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.0.6
djangorestframework==3.13.1
packaging==21.3
Pillow==9.2.0
psycopg==3.0.16
psycopg2==2.9.3
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
sqlparse==0.4.2
tzdata==2022.1
wrapt==1.14.1
